Add config::auth module for auth related code
